+++ R/elo.glm.R
@@ -51,23 +51,23 @@ elo.glm <- function(formula, data, family = "binomial", weights, na.action, subs
 
   # find spanning set
   QR <- qr(dat)
-  dat <- dat[QR$pivot[seq_len(QR$rank)]]
+  dat.qr <- dat[QR$pivot[seq_len(QR$rank)]]
 
-  dat$wins.A <- mf$wins.A
+  dat.qr$wins.A <- mf$wins.A
   grp <- mf$group
 
   wts <- mf$weights
-  dat.glm <- stats::glm(wins.A ~ . - 1, data = dat, family = family, na.action = stats::na.pass, subset = NULL, weights = wts, ...)
-  dat.glm$teams <- all.teams
-  dat.glm$group <- grp
-  dat.glm$elo.terms <- Terms
-  dat.glm$na.action <- stats::na.action(mf)
-  dat.glm$outcome <- attr(mf, "outcome")
+  out <- stats::glm(wins.A ~ . - 1, data = dat.qr, family = family, na.action = stats::na.pass, subset = NULL, weights = wts, ...)
+  out$teams <- all.teams
+  out$group <- grp
+  out$elo.terms <- Terms
+  out$na.action <- stats::na.action(mf)
+  out$outcome <- attr(mf, "outcome")
 
   if(running)
   {
-    dat.mat <- as.matrix(dat[names(dat) != "wins.A"])
-    y <- dat$wins.A
+    dat.mat <- as.matrix(dat)
+    y <- mf$wins.A
 
     ftd <- rep(0, times = nrow(dat))
     grp2 <- group_to_int(grp, skip)
@@ -80,13 +80,15 @@ elo.glm <- function(formula, data, family = "binomial", weights, na.action, subs
       # tmpfit <- stats::glm(wins.A ~ . - 1, data = dat, subset = sbst, weights = wts, family = family)
       # ftd[grp2 == i] <- predict(tmpfit, newdata = dat[grp2 == i, ], type = "link")
 
-      coeff <- stats::glm.fit(dat.mat[sbst, , drop = FALSE], y[sbst], wts[sbst], family = dat.glm$family,
-                              control = dat.glm$control)$coefficients
-      ftd[grp2 == i] <- apply(dat.mat[grp2 == i, , drop = FALSE], 1, function(x) sum(x * coeff, na.rm = TRUE))
+      d <- dat.mat[sbst, , drop = FALSE]
+      coeff <- stats::glm.fit(d, y[sbst], wts[sbst], family = out$family,
+                              control = out$control)$coefficients
+      valid <- colSums(d != 0) > 0
+      ftd[grp2 == i] <- apply(dat.mat[grp2 == i, , drop = FALSE], 1, mult_valid_coef, coeff = coeff, valid = valid)
     }
-    dat.glm$running.values <- dat.glm$family$linkinv(ftd)
-    attr(dat.glm$running.values, "group") <- grp2
+    out$running.values <- out$family$linkinv(ftd)
+    attr(out$running.values, "group") <- grp2
   }
 
-  structure(dat.glm, class = c(if(running) "elo.running", "elo.glm", class(dat.glm)))
+  structure(out, class = c(if(running) "elo.running", "elo.glm", class(out)))
 }