+++ R/elo.winpct.R
@@ -75,7 +75,7 @@ elo.winpct <- function(formula, data, family = "binomial", weights, na.action, s
       dat.tmp$teamA <- dat.tmp$teamA[sbst, , drop = FALSE]
       dat.tmp$teamB <- dat.tmp$teamB[sbst, , drop = FALSE]
 
-      wl <- do.call(eloWinPct, dat)
+      wl <- do.call(eloWinPct, dat.tmp)
       vec <- stats::setNames(wl[[1]], all.teams)
 
       difference <- mean_vec_subset_matrix(vec, dat$teamA+1) - mean_vec_subset_matrix(vec, dat$teamB+1)
@@ -85,7 +85,7 @@ elo.winpct <- function(formula, data, family = "binomial", weights, na.action, s
 
       coeff <- stats::glm.fit(cbind(difference, adj)[sbst, , drop=FALSE],
                               dat.tmp$winsA, family = wl.glm$family, control = wl.glm$control)$coefficients
-      ftd[grp2 == i] <- apply(cbind(difference, adj)[grp2 == i, , drop=FALSE], 1, function(x) sum(x * coeff, na.rm = TRUE))
+      ftd[grp2 == i] <- apply(cbind(difference, adj)[grp2 == i, , drop=FALSE], 1, mult_na_coef, coeff = coeff)
     }
     out$running.values <- wl.glm$family$linkinv(ftd)
     attr(out$running.values, "group") <- grp2